requirements.txt
requests
Pillow
tk